# Conflicts:
#	src/main/java/com/example/controllers/ProductController.java

+++ src/main/java/com/example/repositories/ProductRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ProductRepository extends CrudRepository<Product> {
     List<Product> findByCategory(String category);
-    List<Product> findByPriceBetween(float minPrice, float maxPrice);
+    List<Product> findByPriceBetween(Long minPrice, Long maxPrice);
     public long save(ProductRequest entity);
     void update(Product productRequest);
 }
